Tip measurements are uint16_t
Update BSP.cpp

Update BSP.cpp

+++ source/Core/BSP/BSP_Power.h
@@ -19,9 +19,8 @@ void power_check();
 // Returns the tip resistance in x10 ohms, so 7.5 = 75; 14=140 etc
 uint8_t getTipResistanceX10();
 
-uint8_t getTipThermalMass();
-uint8_t getTipInertia();
-
+uint16_t getTipThermalMass();
+uint16_t getTipInertia();
 
 #ifdef __cplusplus
 }
